kernel: sched: extract thread CPU-usage tracking to usage.h
Move thread CPU usage measurement helpers from ksched.h into a new
kernel/include/usage.h header.
